requirements.txt
streamlit
pytesseract
Pillow
requests
pyttsx3
